mmfood-rag/requirements.txt
datasets==3.1.0
pandas==2.2.2
numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==3.0.1
rapidfuzz==3.9.6
pydantic==2.7.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
